[Update] - Refactor old code - in-progress
[Update] - Refactor old code - in-progress
