+++ .gitignore
@@ -37,3 +37,6 @@ next-env.d.ts
 
 #webstorm
 .idea
+/.idea
+/idea
+```
