+++ components/comment-box-container.tsx
@@ -9,7 +9,7 @@ const CommentBoxContainer = (probs: CommentBoxContainer) => {
     <div className="reply-box-container">
       <hr></hr>
       <span className="comment-profile-circle-line">
-        <span className="comment-profile-circle-img"></span>
+        <span className="comment-profile-circle-img" title="user icons">User icons created by Freepik - Flaticon</span>
       </span>
       <span className="comment-username">{commentusername}</span>
       <span className="comment-post-content">{commentcontent}</span>
