+++ styles/components_style/post-box-container.scss
@@ -78,14 +78,15 @@
         background-clip: padding-box, border-box;
         background-origin: padding-box, border-box;
         overflow: hidden;
+        font-size: 0;
         & > .profile-circle-img {
             position: absolute;
             width: 35px;
             height: 35px;
-            top: 2.5px;
-            left: 2.5px;
+            top: 2px;
+            left: 3px;
             border-radius: 50%;
-            background-image: url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Ft00.deviantart.net%2FSWUCfIShpn7wS8Xn4KIkiiFeAcc%3D%2Ffit-in%2F700x350%2Ffilters%3Afixed_height(100%2C100)%3Aorigin()%2Fpre00%2Fd38b%2Fth%2Fpre%2Ff%2F2018%2F009%2F7%2Fc%2Fdo_you_know_da_wae_by_chander_fox-dbzi5xa.png&f=1&nofb=1&ipt=1996a69800e2ab47e0cd1a3687332c3227daaeef94c0044ddae09ebeca68c0b8&ipo=images");
+            background-image: url("../../assets/images/man.png");
             background-size: contain;
         }
     }
